Dynamic Island: An Interactive User Experience
The Dynamic Island enables new ways to interact with iPhone, featuring a design that blends the line between hardware and software, adapting in real time to show important alerts, notifications, and activities. With the introduction of the Dynamic Island, the TrueDepth camera has been redesigned to take up less of the display area. Without impeding content on the screen, the Dynamic Island maintains an active state to allow users easier access to controls with a simple tap-and-hold. Ongoing background activities like Maps, Music, or a timer remain visible and interactive, and third-party apps in iOS 16 that provide information like sports scores and ride-sharing with Live Activities can take advantage of the Dynamic Island.